requirements.txt
pyaudio
vosk
pynput
pystray
pillow